Introducing IPReflectionAnalysis
Summary:
This diff introduces the interprocedural version of reflection analysis. It utilizes 2 new capabilities I have developed: 1) it is implemented as a preserved analysis pass; 2) uses `SpartaInterprocedural`, the new framework for interprocedural analysis.

This is doing a basic context-insensitive analysis. It's insensitive to an extent that we only use one domain for the parameters for each method. When multiple callsites exist, the caller contexts are joined. This is still sound but may get less accurate result than a context-sensitive analysis.

This is built on top of the existing `ReflectionAnalysis`. I have added new optional arguments that supports the reflection analysis bits. It also doesn't take much to turn this into a context sensitive analysis except that I am unsure how much more time it will take to run and how much value it adds.

Reviewed By: thezhangwei

Differential Revision: D20477097

fbshipit-source-id: d62104d7e7831fbe38b8be7ab73ac274f5de43a1