Decouple the Top state between SingletonDexTypeDomain and SmallSetDexTypeDomain
Summary:
The current behavior of the `DexTypeDomain` is when the Singleton domain joins to top, we also set the SmallSet domain to top.
This is not necessary. The current implementation of the join operation in `SingletonDexTypeDomain` is somewhat limited. We might choose to give up and join to top in some cases. But the information in SmallSet domain is still valid and useful. The change separates the top state between the two domains.

Reviewed By: Feng23

Differential Revision: D21367642

fbshipit-source-id: 8cac1f59a61709a1ed967beb18ec3cd2038f22ab