add a `get` method in `InsertOnlyConcurrentSet`
Summary: This patch adds a `get` method in `InsertOnlyConcurrentSet`, that returns a pointer on the element, or `nullptr` if the element is not in the set. Note that this is not equivalent to `find` because `find` is not thread-safe, but `get` is. This patch also changes `insert` so that it returns a pointer instead of a reference because this is the main use case, and this makes the semantic closer to `std::unordered_set::insert` (which returns an iterator).

Reviewed By: yuxuanchen1997

Differential Revision: D21468629

fbshipit-source-id: 28569b5f5a568825bd613259ca6182508110f8d9