Fix SingleImpl
Summary:
Insert `check-cast` instructions at invokes and field stores, so as to correctly capture requirements of type safety. This may be necessary as interface assignment rules are relaxed in Java (and considered to be enforced at runtime at point-of-use).

For simplicity, insert them in general, and remove the redundant ones afterwards.

Overhead is small. A simple test estimate is smaller than ten percent (from ~30s to ~32s).

Reviewed By: emma0303

Differential Revision: D21417018

fbshipit-source-id: d0d404f1e800ef37fcac80bb4e909948c16d9cc9