Convert FieldTypeAnalyzer to ClinitFieldAnalyzer
Summary:
When using `FieldTypeAnalyzer`, we could potentially propagate incomplete type information for fields to the rest of the program during the global call graph iteration. In this change we limit the field type propagate through local `FieldTypeEnvironment` only to <clinit>s. This is similar to how to handle constant values in IPCP. The change breaks down as the follows:
- Only propagate field type through `WholeProgramState`, except for <clinit>s
- Add `ClinitFieldAnalyzer` to propagate type for static field through local `FieldTypeEnvironment`
- Simplify `FieldTypeAnalyzer` to only update local `FieldTypeEnvironment`
- Extended integration test

Reviewed By: Feng23

Differential Revision: D21421574

fbshipit-source-id: 20e5acf32ea24a983ee27e25363272ce63e0ef48