ir type checker regression test for T64921641
Summary: Currently, the Redex type checker only verifies that for a non-static invoke, the callee object's register is merely an object reference; it doesn't verify that the callee object's register is known to contain a type that is assignable to the callee method reference. This diff is a first step in tightening the type checking around invokes; we now make sure the known type of object is valid w/r/t the method being called. Future work will validate invoke parameters etc.

Reviewed By: agampe

Differential Revision: D21476560

fbshipit-source-id: 5689e06fdbfab82ff786d17572eaa516dc0a00c9