Resolver: properly handle invoke-super
Summary: The callee for invoke-super is unambiguous. However, without knowing where the method ref is being called with an `invoke-super`, we aren't able to correct resolve it. Sometimes the method ref itself is correct, but no guarantee is made here. To correctly resolve the callee for `invoke-super`, we need to know who the caller is. More specifically, the containing class of the caller.

Reviewed By: Feng23

Differential Revision: D21535342

fbshipit-source-id: 7a610860d453bfb63bbff0ecb7316e86d4b274f4