Optionally relax throw tracking requirement when following big-blocks
Summary:
This fixes a bug in the outliner (preventing a crash that I have seen with more aggressive outliner versions higher up in the stack).

The issue is around that the notion of big blocks is dependent on whether adjacent blocks have the same throw edges. This is all fine during the analysis phase, but when the outliner starts mutating code, inserting invokes (to outlined methods) may split blocks, and subtly change which instructions live in a block with throw edges, and that in turn would cut may prematurely cut off big blocks according to  the big blocks iterator.

To address this, i am allowing the big block iterator to ignore throw-edge related cut-offs, which is used during the mutation phase.

Notes:
- Coming up with a small regression test case is hard, and I couldn't, but I did verify that this fixed the issue I was seeing.
- While everything remains correct with this fix, I think that what underlies all this actually a further improvement opportunity: When finding outlinable instructions (that can be covered with the same throws-edges), we can always explore beyond any block covered with throws-edges, as long as those instructions cannot throw. The current big-block cut off that may pull in an entire non-throwing block is just a special case of that.

Reviewed By: minjang

Differential Revision: D21553197

fbshipit-source-id: 130573cb9a007d14d311052c6057f273bdef5415