Fix nullness analysis for instance and static fields
Summary:
We need to collect nullness mapping in the right order. That is we need to distinguish between an explicit null assignment and an implicit one. If a field is not initialized in clinit or ctors, it is implicitly null for the rest of the program. This change collects type and nullness mapping for fields first from clinits and ctors. We apply implicit null assignments if necessary and use that as the starting point when collecting information from the rest of the program.

The change includes the followings:
- FieldTypeAnalyzer -> CtorFieldAnalyzer
- Stop initializing local FieldTypeEnvironment based on global state. (We don't need this information)
- Collect clinits and ctors before the rest of the program
- Update tests

Reviewed By: justinjhendrick

Differential Revision: D21655628

fbshipit-source-id: 8bc95f36d3765fdf70789e63cf59774d455abcde