Expand predicates and matching within match.h
Summary:
Trying to use match.h for testing instruction occurrences showed some holes in expressibility and predicates.
This adds additional predicates on instructions and fields and method ones as well.

Also adding a simpler facility for testing one predicate against a set of instructions, without the tuple and vector of vectors.

Reviewed By: amnn

Differential Revision: D21499871

fbshipit-source-id: 9b6fd95d4edfdcd149ac7079c439b268ca537b23