Split out non-static methods
Summary:
In particular, direct and non-true virtual methods. It's configurable, and the options are enabled by default. Non-static methods are made static when they are relocated to a helper class.
Note that we only relocate static methods when the class has not clinit, but we do relocate non-static methods, as they'd require a constructor (which is never relocated) to run first.
We also gather statics on how many methods are unrelocated, i.e. didn't meet all the requirements, even though their declaring class did.

Reviewed By: emma0303

Differential Revision: D21632148

fbshipit-source-id: 116f3f390e0b589cf20d3cdce503681484d3ff02