Reworking of type analysis
Summary:
This diff enhances and factors out the type analysis that happens in the Outliner.

Background: The purpose of the outliner is to automatically factor out frequently occurring "code snippets" into helper methods, saving space by replacing relatively long code snippets with method invocation instructions. The code snippets are a sequence of IR instructions, and meanwhile the outliner can also capture trees of conditional control-flow. One particular challenge for the outliner is that the outlined helper methods needs to have a proper signature, identifying argument and return types. Selecting those is not entirely straightforward for various reasons:
- Redex does not have a full-fidelity (what matters is matching the Android verifier) type inference / checker engine. Redex' type inference only computes an abstraction of the actual types.
- Even a full-fidelity forward type inference engine would not be sufficient, as not all instruction reveal their types. In particular, `const`-instructions produce a value whose type is becomes material at particular use-sites, deciding whether a constant is to be interpreted as an integer, floating-point value, or (null) object.
- Conversely, not all use-sites impose exact type demands. For example, `if` instructions only mandate that the two incoming values are of the same kind, but doesn't tell us whether they are integer, floating-point value, or objects.
- Thus, depending on the context, the same sequence of IR instructions can be outlined with different type signatures. The purpose of the outliner is to save space, so we want to identify as few permissible type signatures as possible. For that, it is beneficial to prefer identifying the type signatures just by looking a the code snippet to outline, and peek at the surrounding context as little as possible.

This rewrite of the type analysis aims at doing a reasonably good job to deal with these challenges. There's room for improvement; it still may give up sometimes (always playing it safe), and it doesn't always find the minimal set of type signatures for a given code snippet. In any case, the logged statistics indicates that it's doing a much better job than before (`num_candidate_arg_type_not_computed` and `num_candidate_res_type_not_computed` are dramatically reduced, while `num_beneficial_candidates` is significantly improved). The implementation tries to reuse as many of the existing analyses in Redex as possible, combining them in interesting ways, combining "forward" and "backward" analyses as appropriate.

There are three main jobs the type analysis has to solve: 1) Compute a type demand for a register that flows into a code snippet (`get_type_demand`), 2) compute result type for a register that is definately assigned in the code snippet and flows out (`get_result_type`), or `get_inferred_type` in case that a register is only assigned along some control-flow paths, and there's no way around considering the context.

1. Compute a type demand for a register that flows into a code snippet (`get_type_demand`): We look at all instructions in the code snippet that use the register, and narrow all the individual type demands to the most specific one (if any). If this fails (e.g. because some instructions don't tell us enough about their concrete type demands like the `if` instruction, or because after narrowing we are left with more than one type), then we'll consider the context via `get_inferred_type`.

2. Compute result type for a register that flows out (`get_result_type`): We are looking at all assignments to the register that flows out (there can be several in the code of conditional control-flow), and widen what we find. When there are `const`-instructions, we peek (via the existing `ConstantUses` analysis) at how the constant is used in the context, considering a good number of special cases.

3. Compute inferred type case when a register is only assigned along some control-flow paths (`get_inferred_type`): We start using the existing local type inference engine. When it doesn't give us a high-fidelity answer, we take things in our own hands, lookup at the reaching definitions, and then use the same functionality as `get_result_type` to figure out the type.

Another complication that the regular and constant type-demands analysis are dealing with is that certain binary instructions (AND, OR, XOR) can preserve booleanness.

Reviewed By: thezhangwei

Differential Revision: D21513382

fbshipit-source-id: eaeb6378d83da456ecdfae3517ca8f8992091c6a