Outlining of tree-shaped control-flow
Summary:
When reaching a conditional branch or switch instruction, different control-paths are explored as well, as long as they eventually all arrive at a common block. Thus, outline candidates are in fact instruction sequence trees.

A common theme in the code changes is that used to be a simple loop over a sequence of instructions now becomes a walk over the nodes of the tree.

There are also additional statistics on why we give up on a partial candidate.

Reviewed By: justinjhendrick

Differential Revision: D21408375

fbshipit-source-id: 05f8e5fc98423a4cac00981dd5b50b660b9f53be