Add `DexClass::gather_load_types`
Summary:
Add `gather_load_types` to `DexClass`, which collects all the (internal) types necessary to load the given class. These are self, the superclasses and implemented interfaces.

This is different from `gather_types`, which includes methods and fields.

Add a test. Add a helper header to create a simple class hierarchy.

Reviewed By: emma0303

Differential Revision: D21884102

fbshipit-source-id: 8d95fabe45339d0c54bc8b6c221eab94611d3efa