(cp ifield analysis) clear environment if constructor calls a non-static method on itself
Summary: If a constructor set an ifleld, but called an instance method which also happened to set that same field, constprop ifield analysis would not notice, and wrongly propagate what it saw to be a constant value (the initial set of the ifield) to the rest of the constructor. the sfield analysis already takes into consideration clinit calling a static method (and clears the environment), and the same thing needs to happen for ifields w/r/t instance methods.

Reviewed By: ssj933

Differential Revision: D21878232

fbshipit-source-id: 23c7077f89e9e9d5334c0a196ca424e4fc16d5c2