Normalize simple constructors to ease the method deduplication
Summary:
When working on merging anonymous classes with TypeErasure, I find the constructors can not be dedupped because their signatures  and instructions order are different.

This diff figures out the logically equal constructors and only leave one to represent others. Here is an example, the two methods have different signatures but their logic are equal.
```
 * void <init>(B b, A a, D d, C c) {
 *   this.f2 = b;
 *   this.f1 = a;
 *   this.f3 = c;
 *   super.<init>(this, d);
 * }
 *
 * is logically equal to
 *
 * void <init>(A a, B b, C c, D d) {
 *   this.f1 = a;
 *   this.f2 = b;
 *   this.f3 = c;
 *   super.<init>(this, d);
 * }
```

Change the callsites of the second to the first method and adjust the parameters order at callsite. The latter RMU pass can delete the unused one.

Only used in test now, in the later diffs, I'll integrate this to TypeErasure.

Reviewed By: thezhangwei

Differential Revision: D21774978

fbshipit-source-id: 47f8a90bd03f142bb63ff85a5239ba34cf24b925