Resolving callees with call graph
Summary:
The call graph contains information of what callees are possible at each invoke instruction. When performing interprocedural analysis, we should be able to use this information directly instead of relying on method override graph.

When a complete call graph is used, the set of methods returned by this function is equivalent to what a method override graph would return. However, when graphs omit edges (usually for performance reasons), this set might change. In interprocedural analysis, it's important that the set of callees agrees with the call graph it's operating on.

Reviewed By: Feng23, ssj933

Differential Revision: D21947936

fbshipit-source-id: df7ded6906de2b15dfbe2abfd5014c44e1e9a0dc