Improve CopyPropagation monitor handling
Summary:
It is beneficial for the Android verifier if the same lock object is "reached" in the same register, without aliases. As such, try to unify the inputs of `monitor-enter`/`monitor-exit` instructions.

In general this optimization is expected to be a net size benefit, as the introduced `move` should allow at least another one to be eliminated. The impact of the added temp register is not clear. In the best case, the register allocator can coalesce the register.

Reviewed By: NTillmann

Differential Revision: D21912846

fbshipit-source-id: dda86f6cf0de750161654a6f7f1d3b792ecfb9e3