Initial transform for null check removal
Summary:
Based on the nullness result of the analysis, we can evaluate some of the null checks on a reference type value at Redex time.
This change only supports `IF_EQZ` and `IF_NEZ`. We can try to expand to other branch type if there're opportunities left by other passes.

Reviewed By: kuganv

Differential Revision: D22227513

fbshipit-source-id: 1f7953cfa0084a280771dad630ac2165d6149783