APUT_OBJECT only demands object
Summary:
There seems to be very little static verification for this instruction, as most is deferred to runtime.
https://android.googlesource.com/platform/dalvik//android-cts-4.4_r4/vm/analysis/CodeVerify.cpp#186

Also, think of the DEX bytecode that would come out of the following program.
```
Object[] a = new String[1];
Integer o = new Integer(42);;
a[0] = o;
```
after this sequence, a read to `a[0]` MUST NOT be substituted by `o`.

So the formal type demands on an `APUT_OBJECT` are truly `object` for the value, and `object[]` for the array.

Thus, we do the following bug fixes.
- The RemoveArrayLiterals pass should not blindly assume that the values set via APUT_OBJECT on newly created arrays are coming in properly typed for the newly created array-type; instead, the pass will now introduce proper check-cast instructions that properly mimic the semantics of APUT_OBJECT (*) (**)
  - These will get in turn removed by the RemoveRedundantCheckCasts pass where possible
- The CommonSubexpressionElimination pass will no longer blindly forward aput-object to iget-object, as that in fact requires an extra cast to ensure equivalent behavior (which isn't implemented yet, but could be in the future) (**)
- Similarly, the Replace_AputAgetObject peep-hole optimization gets disabled (**)

(*) There's a small change in semantics, actually: The original code using APUT_OBJECT might have thrown a `ArrayStoreException`, while the newly generated code might throw a `ClassCastException` (instead of producing statically invalidly typed DEX code).

These bug fixes get turned into a net size-win, as the RemoveRedundantCheckCasts pass no longer has to dance around the RemoveArrayLiterals pass limitation.

Reviewed By: thezhangwei

Differential Revision: D22173471

fbshipit-source-id: 87939999e28062b0de98c7ff331972800d49fc81