Exclude any init reachables from null check removal
Summary:
There's a known issue in the nullness analysis part. That is in a method invoked by a clinit or ctor a read of a field that is not yet initialized by the 'init' method does not yield the matching nullness result with the result of the analysis. We did see errors since we didn't handle this issue.

This is a simple work around. We gather all methods reachable from a clinit or ctor within the call graph. We put the reachable set into `any_init_reachables`. In the transformation, we do not apply null check removal to methods in the set. We do not perform more complex field value flow analysis, since we don't understand the value of doing that at this point.

Reviewed By: kuganv

Differential Revision: D22405729

fbshipit-source-id: c3e2c83f14f3e1c256a6643544d66b79b4a33516