Don't outline things that involve types/methods/fields incompatible with min-sdk
Summary:
This introduces a new helper class, `RefChecker`, that is useful to check whether a reference (type/method/field) is safe to use in a particular store / min-sdk context.

This is used in this diff to outline less --- we don't want to co-locate outlined code snippets with higher than min-sdk requirements.

In the future, we can use the `RefChecker` when inlining or relocating code in other optimizations.

Reviewed By: Feng23

Differential Revision: D22250215

fbshipit-source-id: ae037c87dd47d017e7ea14900f802fc222bf917c