Avoid gotos or mark as NOLINT
Summary:
Mollify clang-tidy's no-gotos checks. But do not enable the checks as googletest has instances in macros.

To prettify the code and avoid lambdas, change `ControlFlowGraph` to store the original list when not editable (as it cannot be recovered from the iterators easily), and fix `get_param_instructions` for the `m_editable == false` case.

Reviewed By: wsanville

Differential Revision: D22367286

fbshipit-source-id: fc34a50ecf94b4aaa072dc0d19209b72a2ab13f8