Allow removal of unread fields in which only "vestigial" objects are stored
Summary:
We say an object is "vestigial" when the only reference to it is stored in some confined well-known memory locations, not escaping in any other way, and if it only contains vestigial objects in fields or array elements. In other words, the only way to retrieve and observe such an object is by reading from those well-known memory locations.

In this rather simple initial implementation we only consider newly created arrays that are persisted in a single field (of some other object instance), and don't contain any other objects.

Then, if that field is unread, we can drop the field and the iput/sput to it, as it is not possible that the object's lifetime (or even anything stored in that object) can be observed by a weak reference, at least after the storing method returns.

Reviewed By: amnn

Differential Revision: D22565512

fbshipit-source-id: 2de93b12bd6f4c33ef8671f94174f6071173713b