Workaround `whole-archive` issue
Summary:
Autoconf + libtool does not really support whole-archive linking. It is also not portable.

Bite the bullet and have the tests depend on a `libopt` library to share some compilation, while `redex-all` will have to `opt/` directory as sources.

Add a simple tests that checks that the output of `--show-all` is not 0.

Differential Revision: D22669431

fbshipit-source-id: 1c52232c3df753c56d51df9a0ac15e5fe222906a