Workaround `whole-archive` issue (#508)
Summary:
Pull Request resolved: https://github.com/facebook/redex/pull/508

Autoconf + libtool does not really support whole-archive linking. It is also not portable.

Bite the bullet and have the tests depend on a `libopt` library to share some compilation, while `redex-all` will have to `opt/` directory as sources.

Add a simple tests that checks that the output of `--show-all` is not 0.

Reviewed By: thezhangwei

Differential Revision: D22669431

fbshipit-source-id: d640dc9ed274fe69aff05d92b63677e65618496c