Add some MingW cmake directives
Summary:
Specialize some bits for MingW (with the assumption of MSYS2).

`cmake` was chosen as static linking seems to be unhappy under autoconf (pthreads `__imp_` symbols not found). `cmake` is itself not yet completely great, as it does not link in `libwinpthread` and `libstdc++6`, but that is a significantly smaller set than statically-linked autoconf.

Reviewed By: justinjhendrick

Differential Revision: D22780894

fbshipit-source-id: 14296c99372e1ab4b4d73bfeb29a980e235aa67d