Change RedexMappedFile
Summary:
Boost's `mapped_file` does not allow the file to be open on some platforms (even if the other file descriptor is opened for shared mode). As such, ditch the file descriptor and retain the path.

For our purposes, this is equivalent. We do not expect the file to be changed in parallel, so referrring to it by name is as good as with the file descriptor.

Reviewed By: minjang

Differential Revision: D22830757

fbshipit-source-id: 7161a3f279f11db64ae870ed6cace15e0735df69