Remove usage of get_rstate_pure_methods from CSE and reinstate get_immutable_getters
Summary:
CSE uses conditional Purity which is different from commonly known Purity of functions. Thus removing the use from CSE. The pass is now disabled so this is a no-op.

This is needed for correctness. I will explore the usefulness of the purity analysis and remove it if there is no need elsewhere.

Reviewed By: amnn

Differential Revision: D22768303

fbshipit-source-id: 4437644691a58df4359686fe7914b86f9775e879