Improve GraphVisualizer integration
Summary:
Retain unmatched types and try to find them when a new pass is added. This allows to track generated classes.

Add the pass number to the pass name, to distinguish repeated passes. This is useful for the case of generated classes, or in general when a method does not undergo many changes (and thus skips some output). To avoid overhead when no classes are tracked add a lazy name generation overload.

Reviewed By: budmonde

Differential Revision: D22992061

fbshipit-source-id: eff79668f2e43af03d178d5a3e64046ebaaf2bde