No more tuple-based representation
Summary:
Matchers were previously represented by a closure-like type:  A tuple of closed-over variables, and a function pointer that accepted the type being matched and references to all the closed over variables.  This allows matchers to be
- combined without incurring an indirection (e.g. virtual call) cost when they call into each other.
- constrained by the type they are matching on.

But it comes at the cost of a very verbose declaration which has to enumerate the types of the variables being closed over and a sometimes confusing wrapping procedure (which has to "tie the knot" between the tuple and the function pointer).

This diff replaces the representation by a wrapper over a callable object.  The wrapper retains the previous representation's benefits but avoids its pitfalls:
* It is intended to be used with lambdas and uses their captures to close-over any variables.
* The lambda's type then appears as the second template parameter to `match_t`, but cannot be named.  When `match_t` is used in parameter positions, we rely on template deduction in the `matcher` helper function to avoid having to name it.  When `match_t` is used in return type positions, we rely on C++14's return deduction to do the same.

Some other changes I made along the way:

* Addressed some of the TODOs that were simpler to fix than to preserve the old behaviour (especially the ones that came with the promise of free beers).
* Removed unused combinators (especially the very complex or specific ones) -- I didn't want to port these without code that exercised and tested them.
* Moved more implementations into the header for consistency.  Larger implementations remain in the source file, hidden behind a `detail` namespace.

Reviewed By: agampe

Differential Revision: D22776599

fbshipit-source-id: 43b05336216765b123ea2bc39f5ca2b9bab5ec64