Allow matching over non-pointer types.
Summary:
Allow the definition of matching predicates for non-pointer types.  `match_t`
now accepts a const reference to that type, and is specialised for pointers to
take the pointer by value.

This allows us to avoid taking the cost of an extra indirection when defining
predicates on pointers.

This change requires that matchers expecting pointers have a pointer in their
template parameter and matchers that delegate to other matchers match their
signatures with the matchers they delegate to (using the  `arg` struct).

Reviewed By: agampe

Differential Revision: D22817125

fbshipit-source-id: d2814156aa6e8525a19d969e9dbecf3b8150883d