Generalising equals and in predicates.
Summary:
Taking advantage of the flexibility of the new representation to generalise
some core combinators:

`ptr_eq` becomes `equals`.  The former performs checks on pointer types, the
latter performs equality checks on pointers and references.  It is also able to
take ownership of the expected value if it is passed a temporary.

The expected and actual value's types are not different, as C++'s type
deduction isn't powerful enough to figure out the actual value's type, usually.

There is a special case so that pointer types aren't double-indirected.  It was
done this way because the implementation being replaced has this property, but
how important this is is unclear.

Similarly, `in` has been adapted to be able to take ownership of a temporary
container that is passed to it.

Reviewed By: agampe

Differential Revision: D22823222

fbshipit-source-id: 89d08d5e562416df191abb7a37669f708ac796c3