Generate matchers for each opcode predicate
Summary:
Replace ad-hoc definitions for `IRInstruction*` matchers based on the opcode
with matchers that are automatically generated using the opcode drivers.

Created matchers for each oprange, opcode and iopcode.  For the latter two,
as some opcodes conflict with language keywords, I've added an underscore
suffix.  This also helps avoid conflicts with the new `m::invoke_` conflicting
with what is now `m::an_invoke`.

For each range or code, two matchers are created, one that accepts a subsidiary
matcher and one that does not.  The former checks the subsidiary matcher if the
opcode matches.

This diff also removes the `is_opcode` matcher, replacing its invocations with
an equivalent generated matcher.

Reviewed By: agampe

Differential Revision: D23085981

fbshipit-source-id: 33fec255bdc6a88c4c0872d5d6571448f994ce24