Migrating to auto-generated opcode-class predicates.
Summary:
Getting rid of the hand-rolled predicates for ranges of instructions, and
replacing them with invocations of the predicates defined based on `OPRANGE`.

See the previous diff for a rationale for why `_a` or `_an` was added to the
predicate name.

This change also goes some way towards consolidating IROpcode predicates in
IROpcode.h instead of DexInstruction.h

Reviewed By: agampe

Differential Revision: D22849989

fbshipit-source-id: 9477095610fc83d997dbd9c99e4baf7980b634bf