Memoize in purity computation
Summary:
Add a cache to `compute_locations_closure`'s WTO computation. Because of sorting for deterministic behavior, successors can be expensive if repeated over and over.

The cache is very simplistic, as it is "good enough." The cache is limited by the maximum cumulative vector size and does not evict elements. The cache adds entries after a given amount of attempts and if they're above a given threshold size. These knobs were chosen to guide the cache filling so "large" and repetitive methods are likely cached, while "small" data can be recomputed.

By default, cache 4M entries (corresponding to 32MB of direct memory, plus overhead per method). Fill the cache on the second attempt, and only if the vector has at least five items.

The knobs can be twiddled in the JSON config.

Add Trace module `PURITY` to dump cache stats. Add timing for each iteration.

Reviewed By: kuganv

Differential Revision: D23310288

fbshipit-source-id: d59fa5081c11ce0c725392de61845f8926447c2b