Dual layer string map
Summary:
Reduce the amount of string compares by layering the string map.

Previously, the string map was sharded by the `ConcurrentContainer` with the truncated string hash. Now, use an array for the truncated string hash, and the string length for the nested `ConcurrentContainer`s. Thus, strings that do not have the same length will no longer be `strcmp`-ed. This works (without reading the whole string) because we are given the UTF size.

The ordering has been chosen to try and distribute the strings more evenly. Using the string length first would possibly skew the nested containers. The downside of the ordering is that we need key projection, or overhead of storing pairs in the innermost maps.

For a large app, this saves >13% of instructions for the dex loading step, and ~1.5% overall.

Reviewed By: int3, ssj933

Differential Revision: D23437397

fbshipit-source-id: d97cec8bd66c02755d655bed96a568f8cde72b71