fix possible clinit problem in vertical merging
Summary:
Under some situation, a class might not have clinit, but in redex context there exist methodref for its clinit. (Possibly when removing its original clinit the methodref is not cleaned up.) And then that will cause the clinit not being moved to target class because clinit is not being referenced anywhere in code.

Fix this by special handling <clinit> because clinit is not referenced from code.

Reviewed By: Feng23

Differential Revision: D23633141 (https://github.com/facebook/redex/commit/30422c34e2dc812ea62e366ff9334429ae59c1fd)

fbshipit-source-id: 637be698f282086f4ad6fdd73dfc0d4f99256bdb