IRInstruction: Introduce `src_index_t`; unify definition of `param_index_t`
Summary:
We usually use an unsigned integer to refer to one element of `srcs` in an `IRInstruction`. However, it's easy to be confused with `reg_t` as it's easy to forget this layer of indirection. Add this type for future reference.

We use `param_index_t` in a few analyses that index to the method parameters or `srcs` in an invoke instruction. This essentially should be the same type as `src_index_t`.

Reviewed By: minjang

Differential Revision: D23831846

fbshipit-source-id: 545b7312b4bd3867b52d7732a58c0f27d0a85866