LiveRange service: Fix definition of `Use`; Expose def-use and use-def chains analysis
Summary:
Redex has an implementation of computing use-def chains. Expose it in the headers so they can be used in optimization. Its counterpart -- def-use chain can be computed from the same analysis.

Meanwhile there's a problem with current implementation of `Use`. A `Use` was defined as a tuple of (instruction, register), which is not correct for use-def chains, although this does not affect the correctness of `RegAlloc`.

Reviewed By: int3

Differential Revision: D23831845

fbshipit-source-id: 8417e8a719370a71cb44f191f3a912f62218c3a0