fix NPE on inline instrumentation test setup
Summary: Without this JUnit annotation, the setUp method was not being run, which never initialized the lm field. This caused NPEs whenever something tried to execute one of lm's methods. The new line moves the line counts, so I had to update all the strings. Finally, the signature of  `assertEquals()` is `expected`, then `actual`, so I've modified the tests to reflect that.

Reviewed By: wsanville

Differential Revision: D23784010

fbshipit-source-id: aa2acc2b020cd7b38d69a1eb9fc8755eb8b54bb8