Add IODI layers
Summary:
Add "layers" to IODI. Layers are separate line ranges to disambiguate overloads. Use type "`iodi2`" for the layered encoding for now.

Layers are encoded in the top bits of the 32-bit line numbers. Currently reserve the top four bits, restricting "normal" line ranges to be below ~268M, which seems acceptable. Do not use the top bit to avoid any issues, as lines are interpreted as signed on the Java side.

Layer encoding extends the "base" encoding: if a debug item's line start without the layer bits is `0`, an IODI program is in use. Otherwise, it is a direct encoding of the source line numbers.

Computation is heuristical as an iteration over the layers:
* For each layer, parameter-size sets with "big" overloads are attempted first. If an overload is handled, other overloads in the same cluster are skipped to the next iteration.
 The downside of this approach is suboptimality if the "big" overload is too large for IODI. A prototype shows that the added complexity only achieves so much (~3K over ~157K for a large app), as a result of which the complexity is left to a follow-up.

* For each parameter-size set, overloads with the same size form a search space. As combinatorial explosion is pretty bad, only use a very simple heuristic (single "parallel" traversal of the clusters) to find the best local solution. This works acceptably well.

Small future savings are possible for corner cases. For example, if a cluster's methods are always using the direct encoding, layering only adds overhead. Lazy writing and a post-pass could do better.

A change to the metadata encoding is left to a follow-up diff.

Add tests (where possible).

Reviewed By: newobj

Differential Revision: D23922850

fbshipit-source-id: ac14a99c27c19f36165f99630f30cdf59ab638ee