Include potential call backs in any_init_reachables
Summary:
We do not apply null check removal on methods reachable from any init methods. The reason is that the nullness status is unknown before the init method finishes. However, this is not enough.

We discovered that a vmethod on an anonymous class can be implicitly reachable from a ctor through a call back from another thread. That is a call back could occur by another thread while the ctor of the enclosing class is not finished yet.

To cover this tricky case, we include a `track_callbacks` flag while scanning the `any_init_reachable` set. If the flag is true we include all vmethods of the anonymous class to our scan.

Reviewed By: ssj933

Differential Revision: D24072026

fbshipit-source-id: 5fbf3de94cc110b744bf77e4ce14c8f5b2db13b6