Use `std::ref` for `dexmethods_profiled_comparator`
Summary:
Try to avoid new issues with stateful comparators being expensive with `libstdc++`. Wrap the comparator with `std::ref` in `DexOutput`, and add more warning messages. Refactor the cache back into the comparator.

We could add a wrapper that works with a `shared_ptr`. However, that adds overhead (indirection, destructor checking), which seems complexity for no further gain (for this specific case).

Reviewed By: Feng23

Differential Revision: D24262202

fbshipit-source-id: 0041f8abccbe8b475f8ca2ecc63ba8d1840b1409