Remove IODI nondeterminism
Summary:
The order of methods in a param-size set may determine which methods will use IODI. Previously we seem to have gotten lucky that the simple method pointer, just to distinguish things, was acceptable.

The complexities of IODIv2 expose this ordering nondeterminism, as different "punt to next layers" may happen.

Reviewed By: wsanville

Differential Revision: D24434763

fbshipit-source-id: 94075cebb45b59fe7ca4b5053e09c17c7aab822c