Fix NullnessDomain propagation for primitive arrays
Summary: When analyzing an `aget-primitive`, we forgot to overwrite the destination register (most likely with Top). This mis-handling causes the analyzer to pick up the nullness domain of the array object as the domain for the array element, and propagates it downwards. The incorrect nullness info then causes the nullness check removal to remove the wrong `IF_EQZ`.

Reviewed By: yuxuanchen1997

Differential Revision: D24559951

fbshipit-source-id: 0ab68e044a1feb2a43a17273ca763bf36500906f