Minor code refactoring and MergingStrategies introducing
Summary:
The existing class merging strategy is to split the mergeables into several groups by considering the configured min and max count. The diff refactors the logic out to a separate file, so that we could be able to add different merging strategies latter.

The existing merging strategy has a hack of adding the trailing class to the former group if only one class left. It's removed by the diff and the bsb might show minor changes.

Reviewed By: thezhangwei, minjang

Differential Revision: D24582059

fbshipit-source-id: a84161285e73226b4dbe9ca1fae4d01c5553ef05