Add a method to derive sizes after passes
Summary:
Add the capability to write temporary output after each pass to derive the "current output size," and write the value into the pass metrics.

The approach here is to `fork()` the process after each pass, when it is (assumed to be) single-threaded and it is defined behavior to not `exec()`. This way the "main" process can make regular progress.

In the child, run `MakePublicPass` and `RegAllocPass` as possibly necessary fixups (though the latter is quite expensive), but no further passes. Change the output directory to an agreed-upon temporary directory, and finish normally.

In the parent, keep a list of such child processes and regularly check whether some are done. In that case, sum up the dex file sizes in the according temporary directory, write the stat, and clean up.

As Redex can be quite memory-hungry, add an option that restricts the backlog of child processes. If the queue is full, wait till at least one child finishes.

Sample invocation:
```
$ redex.py -J after_pass_size=true -J after_pass_size_queue=4 [...]
```

Reviewed By: Feng23

Differential Revision: D24660475

fbshipit-source-id: ef530c2ebe827646015c74c0765c2cf7b829b175